requirements.txt
flask
pymysql
sqlalchemy
python-dotenv